refactor MD editor to components folder
